access control

+++ pages/delete_project.php
@@ -1,7 +1,7 @@
 <?php
-if (isset($_GET['project_id'])) {
+if (isset($_GET['project_id']) && User::getCurrentUser()) {
     $project = Project::getProjectById($_GET['project_id']);
-    if ($project) {
+    if ($project && ($project->owner_id == User::getCurrentUser()->id || User::getCurrentUser()->is_admin)) {
         $project->delete();
     }
     redirect(url(['_page' => 'home']));
